Conflicts:
	Uebung5/skeleton-cg1_ex4 - Friedrich2/Texture.cpp

+++ Uebung5/skeleton-cg1_ex4 - Friedrich2/Texture.hpp
@@ -64,5 +64,5 @@ namespace World{
 
 
   void raytrace(int x, int y);
-  Ray getRay(int x, int y, int xPixel, int yPixel);
+  Ray getRay(int x, int y, float xPixel, float yPixel);
 };
